added database connection string and adding a Volunteer into the database

+++ src/main/java/com/example/volunteerapp/Volunteer.java
@@ -2,11 +2,13 @@ package com.example.volunteerapp;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.security.SecureRandom;
+import java.sql.*;
 import java.time.LocalDate;
 import java.time.Period;
 
@@ -24,7 +26,7 @@ public class Volunteer {
         // setImageFile(new File(Volunteer.class.getResource("/com/example/volunteerapp/images/defaultPerson.png").getFile()));
     }
 
-    public Volunteer(String firstName, String lastName, String phoneNumber, LocalDate birthDate, File imageFile) throws IOException {
+    public Volunteer(String firstName, String lastName, String phoneNumber, LocalDate birthDate, File imageFile) throws IOException, URISyntaxException {
         this(firstName, lastName, phoneNumber, birthDate); // it will call the constructor with default image file and then will override it in the next line
         setImageFile(imageFile);
         copyImageFile();
@@ -92,27 +94,33 @@ public class Volunteer {
     /**
      * This method will copy the file specified to the image directory on this server and give it
      */
-    public void copyImageFile() throws IOException {
+    public void copyImageFile() throws IOException, URISyntaxException {
         // create a new Path to the image into a local directory
         Path sourcePath = imageFile.toPath();
 
         String uniqueFileName = getUniqueFileName(imageFile.getName());
+        // professor Jaret's code
+        Path targetPath = Paths.get("src/main/resources/com/example/volunteerapp/images/" + uniqueFileName);
 
-        Path targetPath = Paths.get(Volunteer.class.getResourceAsStream("images/" + uniqueFileName).toString());
+        // Path targetPath = Paths.get(Volunteer.class.getResourceAsStream("images/" + uniqueFileName).toString());
         // Path targetPath = Paths.get(Volunteer.class.getResource("/com/example/volunteerapp/images/" + uniqueFileName).getFile());
 
-        // copy the file to the new directory
-        Files.copy(sourcePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
+        // more of professor's code
+        File fileSource = new File(sourcePath.toString());
+        if(fileSource.exists()){
+            // copy the file to the new directory
+            Files.copy(sourcePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
 
-        // update the imageFile to point to new File
-        imageFile = new File(targetPath.toString());
+            // update the imageFile to point to new File
+            imageFile = new File(targetPath.toString());
+        }
     }
 
     /**
      * This method will recieve a String thet represents a file name and return a String with a random,
      * unique set of letters prefixed to it.
      */
-    private String getUniqueFileName(String oldFileName){
+    private String getUniqueFileName(String oldFileName) throws URISyntaxException {
         String newName;
 
         //create a Random Number Generator (rng)
@@ -143,15 +151,22 @@ public class Volunteer {
      * This method will search the iamge directoy and ensure that the file name
      * is unique
      */
-    public boolean uniqueFileInDirectory(String fileName){
-        File directory = new File(Volunteer.class.getResourceAsStream("images/").toString());
+    public boolean uniqueFileInDirectory(String fileName) throws URISyntaxException {
+        // File directory = new File(Volunteer.class.getResourceAsStream("images/").toString());
         // File directory = new File(Volunteer.class.getResource("/com/example/volunteerapp/images/").getFile());
 
-        File[] dir_contents = directory.listFiles();
+        // Professor Jaret's code
+        String path = Volunteer.class.getResource("images/").toURI().getPath();
+        File directory = new File(path);
 
-        for(File file : dir_contents){
-            if(file.getName().equals(fileName)){
-                return false;
+        // more of professor's code
+        if(directory.isDirectory()){
+            File[] dir_contents = directory.listFiles();
+
+            for(File file : dir_contents){
+                if(file.getName().equals(fileName)){
+                    return false;
+                }
             }
         }
         return true;
@@ -186,5 +201,43 @@ public class Volunteer {
      */
     public String toString(){
         return String.format("%s %s is %d years old", firstName, lastName, Period.between(birthDate, LocalDate.now()).getYears());
+    }
+
+    /**
+     * This method will write the instance of the Volunteer into the database
+     */
+    public void insertIntoDB() throws SQLException {
+        Connection conn = null;
+        PreparedStatement preparedStatement = null;
+
+        try{
+            //1. Connect to the database
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/volunteer", "root", "password");
+            //2. Create a String that holds the query with ? as a user inputs
+            String sql = "INSERT INTO volunteers (firstName, lastName, phoneNumber, birthDate, imageFile) " +
+                    "Values (?,?,?,?,?)";
+            //3. prepare the query
+            preparedStatement = conn.prepareStatement(sql);
+            //4. Convert the birthdate to a SQL date
+            Date db = Date.valueOf(birthDate);
+            //5. Bind the values to the parameters
+            preparedStatement.setString(1, firstName);
+            preparedStatement.setString(2, lastName);
+            preparedStatement.setString(3, phoneNumber);
+            preparedStatement.setDate(4, db); // birthDate
+            preparedStatement.setString(5, imageFile.getName()); // name of the image file
+
+            preparedStatement.executeUpdate();
+        } catch (Exception e){
+            System.err.println(e.getMessage());
+        } finally {
+            if(preparedStatement != null){
+                preparedStatement.close();
+            }
+
+            if(conn != null){
+                conn.close();
+            }
+        }
     }
 }
